Fixed webhook messages edition via references.

+++ cogs/function_webhook/lecture_webhook.py
@@ -1,17 +1,16 @@
 import re
 import sqlite3
-import urllib.request
 
 import discord
 
 
-async def edit_webhook(message: discord.Message, idS, user):
+async def edit_webhook(bot, message: discord.Message, idS, user):
     db = sqlite3.connect("src/owlly.db", timeout=3000)
     c = db.cursor()
     ref_id = message.reference.message_id
-    ref_msg = await message.channel.fetch_message(ref_id)
+    ref_msg = message.reference.resolved
     perso_name = ref_msg.author.name
-    sql = "SELECT idDC WHERE (idS = ? AND idU=? AND Nom = ?)"
+    sql = "SELECT idDC FROM DC WHERE (idS = ? AND idU=? AND Nom = ?)"
     var = (
         idS,
         user,
@@ -19,13 +18,11 @@ async def edit_webhook(message: discord.Message, idS, user):
     )
     c.execute(sql, var)
     perso_check = c.fetchone()
-    if perso_check:
-        id_edit = ref_msg.webhook_id
-        if isinstance(id_edit, int):
-            webhook = await message.channel.fetch_webhook(id_edit)
-            await webhook.edit_message(id=ref_id, content=message.content)
-            await message.delete()
-            return
+    if perso_check and isinstance(ref_msg, discord.Message):
+        webhook = await get_bot_webhook(bot, message.channel)
+        await webhook.edit_message(message_id=ref_id, content=message.content)
+        await message.delete()
+        return
     c.close()
     db.close()
 
@@ -43,6 +40,26 @@ async def delete_HRP(message: discord.Message, idS):
         return
     c.close()
     db.close()
+
+async def get_bot_webhook(bot, channel : discord.TextChannel):
+    """Returns the Owlly webhook used for personae in a text channel.
+    If there isn't one, it creates the webhook and returns it."""
+    channel_webhooks = await channel.webhooks()
+    OwllyNPC = None
+    if len(channel_webhooks) :
+        # Webhook exists
+        for i in channel_webhooks:
+            if i.user == bot.user:
+                # A Webhook created by the bot exists
+                OwllyNPC = i
+                break
+    if not OwllyNPC:
+        OwllyNPC = await channel.create_webhook(
+            name="OwllyNPC",
+            avatar=(await bot.user.avatar_url.read()),
+            reason="OwllyNPC doesn't exist yet !",
+        )
+    return OwllyNPC
 
 
 async def switch_persona(bot, message: discord.Message, idS, user):
@@ -100,22 +117,7 @@ async def switch_persona(bot, message: discord.Message, idS, user):
             web = True
             char = [x for x in char]
     if web:
-        NPC = await message.channel.webhooks()
-        OwllyE = False
-        if len(NPC) > 0:
-            # Webhook exists
-            for i in NPC:
-                if i.name == "OwllyNPC":
-                    # OWLLYNPC exists
-                    OwllyNPC = i  # Est un webhook
-                    OwllyE = True
-                    break
-        if not OwllyE:
-            OwllyNPC = await message.channel.create_webhook(
-                name="OwllyNPC",
-                avatar=(await bot.user.avatar_url.read()),
-                reason="OwllyNPC doesn't exist yet !",
-            )
+        OwllyNPC = await get_bot_webhook(bot, message.channel)
         await OwllyNPC.send(
             content=webcontent,
             username=char[0],
